Refactor to use service functions for fetching bands, variables, statistics, and histogram data
